Rename functions to hooks
