+++ hooks/developertools.go
@@ -1,4 +1,4 @@
-package main
+package hooks
 
 import (
 	"fmt"
