+++ lib/state/BUILD.bazel
@@ -2,10 +2,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "go_default_library",
-    srcs = [
-        "builddata.go",
-        "main.go",
-    ],
+    srcs = ["main.go"],
     importpath = "github.com/nugget/phoebot/lib/state",
     visibility = ["//visibility:public"],
     deps = [
